@@ -50,9 +50,11 @@ dependencies:
 
   geolocator: ^9.0.2
   kakao_map_plugin: ^0.3.1
-  webview_flutter: ^4.0.7
+  webview_flutter: ^4.8.0
   http: ^1.2.2
   freezed_annotation: ^2.4.4
+  tosspayments_widget_sdk_flutter: ^1.0.3
+  url_launcher: ^6.3.0
 
 
 
