@@ -2,7 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:hankkitoktok/const/color.dart';
 import 'package:hankkitoktok/const/style2.dart';
 
-void showWarningDialog(BuildContext context, String title, String content, String leftbutton, String rightbutton, VoidCallback  rightButtonFunction) {
+
+void showWarningDialog(BuildContext context, String title, String content, String leftbutton, String rightbutton, Function? leftButtonFunction, Function? rightButtonFunction) {
+
   showDialog(
     context: context,
     builder: (BuildContext context) {
@@ -33,6 +35,7 @@ void showWarningDialog(BuildContext context, String title, String content, Strin
           TextButton(
             onPressed: () {
               Navigator.of(context).pop();
+              if(leftButtonFunction != null) leftButtonFunction();
             },
             child: Text(
               '${leftbutton}',
@@ -52,6 +55,7 @@ void showWarningDialog(BuildContext context, String title, String content, Strin
             onPressed: () {
               rightButtonFunction();
               Navigator.of(context).pop();
+              if(rightButtonFunction != null) rightButtonFunction();
             },
             child: Text(
               '${rightbutton}',
