@@ -1,6 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
 import 'package:get/get_rx/get_rx.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 import '../models/address/address.dart';
 import '../models/address/address_data.dart';
@@ -12,6 +13,7 @@ import '../models/user/user_data.dart';
 //실행되는 시점: 앱 시작 시
 class AddressController extends GetxController{
   List<Address> addresses = [];
+  Address selectedAddress = Address.init();
   int selectedAddressIndex = 0;
 
   //컨트롤러가 Put 되는 시점에 주소 정보 가져오기(앱 시작시 해당 생성자 실행)
@@ -19,12 +21,32 @@ class AddressController extends GetxController{
   void onInit() async {
     // TODO: implement onInit
     await initAddresses();
+    for(var address in addresses){
+      if(address.addressStatus == AddressStatus.CONFIGURED){
+        selectedAddress = address;
+      }
+    }
     super.onInit();
   }
 
   Future<void> initAddresses()async {
     debugPrint('initAddresses');
     addresses = await addressGetList();
+    orgAddressList();
+  }
+
+  void orgAddressList(){
+    int idx = 0;
+    List<Address> notConfiguredAddress = [];
+    for(var address in addresses){
+      if(address.addressStatus != AddressStatus.CONFIGURED){
+        notConfiguredAddress.add(address);
+      }else{
+        selectedAddress = address;
+        selectedAddressIndex = idx;
+      }
+    }
+    addresses = notConfiguredAddress;
     update();
   }
 
@@ -34,12 +56,8 @@ class AddressController extends GetxController{
   }
 
   void setSelectedAddressIndex(int index){
-    selectedAddressIndex = index;
+    //기존 선택된 주소를 보이게 하고 새로 선택된 주소를 안보이게 함
     update();
-  }
-
-  Address get getSelectedAddress{
-    return addresses.isNotEmpty ? addresses[selectedAddressIndex] : Address.init();
   }
 
   List<String> get getAddressList{
@@ -48,6 +66,19 @@ class AddressController extends GetxController{
       addressList.add(address.address);
     }
     return addressList;
+  }
+
+  Future<bool> setConfiguredAddress(int addressId)async{
+
+    bool result = await networkPatchAddress(addressId);
+
+    if(result) await initAddresses();
+    return result;
+  }
+
+  Future<void> deleteAddress(int addressId)async{
+    bool result = await networkDeleteAddress(addressId);
+    if(result) await initAddresses();
   }
 
 }
