@@ -64,7 +64,7 @@
 //     DateTime.now());
 
 import 'package:hankkitoktok/models/meal/meal_delivery.dart';
-import 'package:hankkitoktok/models/meal/ordered_meal.dart';
+import 'package:hankkitoktok/models/meal/meal_delivery_order.dart';
 import 'package:hankkitoktok/models/meal/meal.dart';
 import 'package:hankkitoktok/models/order/order_post.dart';
 import 'package:hankkitoktok/models/sidedish/sidedish.dart';
@@ -135,158 +135,158 @@ Meal getMealById(int mealId) {
   return mealMenuList[1];
 }
 
+//
+// List<OrderedMeal> orderedMeals = [
+//   OrderedMeal.init(
+//     mealId: 1,
+//     reservedDate: DateTime(2024, 8, 15), // 특정 날짜로 설정
+//     reservedTime: Time.AFTERNOON, // 점심 시간
+//     includeRice: true, // 밥 포함
+//     hasFullDiningOption: false, // 전체 다이닝 옵션 제외
+//   ),
+//   OrderedMeal.init(
+//     mealId: 2,
+//     reservedDate: DateTime(2024, 8, 16),
+//     reservedTime: Time.EVENING, // 저녁 시간
+//     includeRice: false,
+//     hasFullDiningOption: true,
+//   ),
+//   OrderedMeal.init(
+//     mealId: 3,
+//     reservedDate: DateTime(2024, 8, 17),
+//     reservedTime: Time.AFTERNOON,
+//     includeRice: true,
+//     hasFullDiningOption: true,
+//   ),
+//   OrderedMeal.init(
+//     mealId: 4,
+//     reservedDate: DateTime(2024, 8, 17),
+//     reservedTime: Time.EVENING,
+//     includeRice: false,
+//     hasFullDiningOption: false,
+//   ),
+// ];
+//
 
-List<OrderedMeal> orderedMeals = [
-  OrderedMeal.init(
-    mealId: 1,
-    reservedDate: DateTime(2024, 8, 15), // 특정 날짜로 설정
-    reservedTime: Time.AFTERNOON, // 점심 시간
-    includeRice: true, // 밥 포함
-    hasFullDiningOption: false, // 전체 다이닝 옵션 제외
-  ),
-  OrderedMeal.init(
-    mealId: 2,
-    reservedDate: DateTime(2024, 8, 16),
-    reservedTime: Time.EVENING, // 저녁 시간
-    includeRice: false,
-    hasFullDiningOption: true,
-  ),
-  OrderedMeal.init(
-    mealId: 3,
-    reservedDate: DateTime(2024, 8, 17),
-    reservedTime: Time.AFTERNOON,
-    includeRice: true,
-    hasFullDiningOption: true,
-  ),
-  OrderedMeal.init(
-    mealId: 4,
-    reservedDate: DateTime(2024, 8, 17),
-    reservedTime: Time.EVENING,
-    includeRice: false,
-    hasFullDiningOption: false,
-  ),
-];
+// List<MealDelivery> mealDeliveries = [
+//   MealDelivery.init(
+//     mealDeliveryId: _requestIdCounter++,
+//     orderId: 1,
+//     orderedMeal: orderedMeals[0],
+//     orderState: OrderState.DELIVERING,
+//     deliveryState: DeliveryState.DELIVERED,
+//     deliveryRequestTime: DateTime(2024, 8, 15, 12, 30),
+//     orderTime: DateTime(2024, 8, 14, 14, 20),
+//     deliveryStartTime: DateTime(2024, 8, 15, 12, 30),
+//     deliveryCompleteTime: null,
+//   ),
+//   MealDelivery.init(
+//     mealDeliveryId: _requestIdCounter++,
+//     orderId: 2,
+//     orderedMeal: orderedMeals[1],
+//     orderState: OrderState.PENDING,
+//     deliveryState: DeliveryState.INDELIVERING,
+//     deliveryRequestTime: DateTime(2024, 8, 15, 12, 30),
+//     orderTime: DateTime(2024, 8, 14, 14, 20),
+//     deliveryStartTime: null,
+//     deliveryCompleteTime: null,
+//   ),
+//   MealDelivery.init(
+//     mealDeliveryId: _requestIdCounter++,
+//     orderId: 3,
+//     orderedMeal: orderedMeals[2],
+//     orderState: OrderState.DELIVERING,
+//     deliveryState: DeliveryState.PENDING,
+//     deliveryRequestTime: DateTime(2024, 8, 15, 12, 30),
+//     orderTime: DateTime(2024, 8, 14, 14, 20),
+//     deliveryStartTime: DateTime(2024, 8, 17, 12, 30),
+//     deliveryCompleteTime: null,
+//   ),
+//   MealDelivery.init(
+//     mealDeliveryId: _requestIdCounter++,
+//     orderId: 4,
+//     orderedMeal: orderedMeals[3],
+//     orderState: OrderState.ORDERED,
+//     deliveryState: DeliveryState.PENDING,
+//     deliveryRequestTime: DateTime(2024, 8, 15, 12, 30),
+//     orderTime: DateTime(2024, 8, 14, 14, 20),
+//     deliveryStartTime: null,
+//     deliveryCompleteTime: null,
+//   ),
+// ];
+//
+// MealDelivery getMealDeliveryById(int mealDeliveryId) {
+//   for(MealDelivery mealDelivery in mealDeliveries){
+//     if(mealDelivery.mealDeliveryId == mealDeliveryId){
+//       return mealDelivery;
+//     }
+//   }
+//   return mealDeliveries[0];
+// }
+//
+// MealDelivery getNextMealDelivery(String orderId) {
+//   return mealDeliveries[1];
+// }
+//
+// List<List<MealDelivery>> orderedMealsExamples = [
+//   [mealDeliveries[0]],
+//   [mealDeliveries[1], mealDeliveries[2]],
+//   [mealDeliveries[3], mealDeliveries[0], mealDeliveries[1]],
+//   [mealDeliveries[2], mealDeliveries[3], mealDeliveries[0], mealDeliveries[1]],
+// ];
 
+// Order exampleOrder = Order.init(
+//   orderID: _orderIdCounter++,
+//   orderType: OrderType.SCHEDULED, // 주간 결제로 설정
+//   orderState: OrderState.DELIVERING, // 주문 상태 설정
+//   specialInstruction: '배송 시 문 앞에 두세요.', // 요청 사항
+//   userId: 1,
+//
+//   mealPrice: 18000,
+//   deliveryPrice: 3000,
+//   fullServicePrice: 5000,
+//   totalPrice: 26000,
+//   orderTime: DateTime(2024, 8, 14, 14, 20), // 주문 시간 설정
+// );
+//
+// Order exampleOrder2 = Order.init(
+//   orderID: _orderIdCounter++,
+//   orderType: OrderType.IMMEDIATE,
+//   orderState: OrderState.PENDING,
+//   specialInstruction: '배송 시 연락주세요.',
+//   userId: 1,
+//
+//   mealPrice: 12000,
+//   deliveryPrice: 2000,
+//   fullServicePrice: 3000,
+//   totalPrice: 17000,
+//   orderTime: DateTime(2024, 8, 15, 12, 30),
+// );
+//
+// Order exampleOrder3 = Order.init(
+//   orderID: _orderIdCounter++,
+//   orderType: OrderType.SCHEDULED,
+//   orderState: OrderState.DELIVERED,
+//   specialInstruction: '부재 시 경비실에 맡겨주세요.',
+//   userId: 1,
+//
+//   mealPrice: 24000,
+//   deliveryPrice: 4000,
+//   fullServicePrice: 7000,
+//   totalPrice: 35000,
+//   orderTime: DateTime(2024, 8, 16, 11, 45),
+// );
 
-List<MealDelivery> mealDeliveries = [
-  MealDelivery.init(
-    mealDeliveryId: _requestIdCounter++,
-    orderId: 1,
-    orderedMeal: orderedMeals[0],
-    orderState: OrderState.DELIVERING,
-    deliveryState: DeliveryState.DELIVERED,
-    deliveryRequestTime: DateTime(2024, 8, 15, 12, 30),
-    orderTime: DateTime(2024, 8, 14, 14, 20),
-    deliveryStartTime: DateTime(2024, 8, 15, 12, 30),
-    deliveryCompleteTime: null,
-  ),
-  MealDelivery.init(
-    mealDeliveryId: _requestIdCounter++,
-    orderId: 2,
-    orderedMeal: orderedMeals[1],
-    orderState: OrderState.PENDING,
-    deliveryState: DeliveryState.INDELIVERING,
-    deliveryRequestTime: DateTime(2024, 8, 15, 12, 30),
-    orderTime: DateTime(2024, 8, 14, 14, 20),
-    deliveryStartTime: null,
-    deliveryCompleteTime: null,
-  ),
-  MealDelivery.init(
-    mealDeliveryId: _requestIdCounter++,
-    orderId: 3,
-    orderedMeal: orderedMeals[2],
-    orderState: OrderState.DELIVERING,
-    deliveryState: DeliveryState.PENDING,
-    deliveryRequestTime: DateTime(2024, 8, 15, 12, 30),
-    orderTime: DateTime(2024, 8, 14, 14, 20),
-    deliveryStartTime: DateTime(2024, 8, 17, 12, 30),
-    deliveryCompleteTime: null,
-  ),
-  MealDelivery.init(
-    mealDeliveryId: _requestIdCounter++,
-    orderId: 4,
-    orderedMeal: orderedMeals[3],
-    orderState: OrderState.ORDERED,
-    deliveryState: DeliveryState.PENDING,
-    deliveryRequestTime: DateTime(2024, 8, 15, 12, 30),
-    orderTime: DateTime(2024, 8, 14, 14, 20),
-    deliveryStartTime: null,
-    deliveryCompleteTime: null,
-  ),
-];
+//List<Order> orders = [exampleOrder, exampleOrder2, exampleOrder3];
 
-MealDelivery getMealDeliveryById(int mealDeliveryId) {
-  for(MealDelivery mealDelivery in mealDeliveries){
-    if(mealDelivery.mealDeliveryId == mealDeliveryId){
-      return mealDelivery;
-    }
-  }
-  return mealDeliveries[0];
-}
-
-MealDelivery getNextMealDelivery(int orderId) {
-  return mealDeliveries[1];
-}
-
-List<List<MealDelivery>> orderedMealsExamples = [
-  [mealDeliveries[0]],
-  [mealDeliveries[1], mealDeliveries[2]],
-  [mealDeliveries[3], mealDeliveries[0], mealDeliveries[1]],
-  [mealDeliveries[2], mealDeliveries[3], mealDeliveries[0], mealDeliveries[1]],
-];
-
-Order exampleOrder = Order.init(
-  orderID: _orderIdCounter++,
-  orderType: OrderType.SCHEDULED, // 주간 결제로 설정
-  orderState: OrderState.DELIVERING, // 주문 상태 설정
-  specialInstruction: '배송 시 문 앞에 두세요.', // 요청 사항
-  userId: 1,
-
-  mealPrice: 18000,
-  deliveryPrice: 3000,
-  fullServicePrice: 5000,
-  totalPrice: 26000,
-  orderTime: DateTime(2024, 8, 14, 14, 20), // 주문 시간 설정
-);
-
-Order exampleOrder2 = Order.init(
-  orderID: _orderIdCounter++,
-  orderType: OrderType.IMMEDIATE,
-  orderState: OrderState.PENDING,
-  specialInstruction: '배송 시 연락주세요.',
-  userId: 1,
-
-  mealPrice: 12000,
-  deliveryPrice: 2000,
-  fullServicePrice: 3000,
-  totalPrice: 17000,
-  orderTime: DateTime(2024, 8, 15, 12, 30),
-);
-
-Order exampleOrder3 = Order.init(
-  orderID: _orderIdCounter++,
-  orderType: OrderType.SCHEDULED,
-  orderState: OrderState.DELIVERED,
-  specialInstruction: '부재 시 경비실에 맡겨주세요.',
-  userId: 1,
-
-  mealPrice: 24000,
-  deliveryPrice: 4000,
-  fullServicePrice: 7000,
-  totalPrice: 35000,
-  orderTime: DateTime(2024, 8, 16, 11, 45),
-);
-
-List<Order> orders = [exampleOrder, exampleOrder2, exampleOrder3];
-
-Order getOrderById(int orderId) {
-  for(Order order in orders){
-    if(order.orderID == orderId){
-      return order;
-    }
-  }
-  return orders[0];
-}
+// Order getOrderById(int orderId) {
+//   for(Order order in orders){
+//     if(order.orderID == orderId){
+//       return order;
+//     }
+//   }
+//   return orders[0];
+// }
 
 User exampleUser = User.init(
   userId: 1,
