@@ -6,7 +6,8 @@ import 'package:get/get.dart';
 import 'package:hankkitoktok/component/calendar.dart';
 
 import 'package:hankkitoktok/controller/delivery_controller.dart';
-import 'package:hankkitoktok/controller/list_view_scroll_controller.dart';
+
+import 'package:hankkitoktok/controller/history_controller.dart';
 
 import 'package:hankkitoktok/controller/ordered_meal_controller.dart';
 import 'package:hankkitoktok/controller/user_controller.dart';
@@ -22,7 +23,9 @@ import 'package:hankkitoktok/screen/3_menu_choice/1_choice_menu_screen.dart';
 import 'package:hankkitoktok/screen/3_menu_choice/2_order_screen.dart';
 import 'package:hankkitoktok/screen/4_my_page/0_my_page_home.dart';
 import 'package:hankkitoktok/screen/4_my_page/3_delete_id.dart';
+//import 'package:hankkitoktok/screen/4_pay_choice/payment_test.dart';
 import 'package:hankkitoktok/screen/5_order/1_full_dining_explanation.dart';
+import 'package:hankkitoktok/screen/test_screen.dart';
 import 'package:intl/intl.dart';
 import 'package:intl/date_symbol_data_local.dart';
 
@@ -119,8 +122,8 @@ void _notificationSetting(){
 
 void _getControllerSetting(){
   Get.put(OrderedMealController());
-
   Get.put(ListViewScrollController());
+  Get.put(HistoryController());
   Get.put(UserController());
   Get.put(AddressController());
   Get.put(DeliveryController());
@@ -199,12 +202,12 @@ class MyApp extends StatelessWidget {
         //   page: () => CommunityScreen(),
         // ),
       ],
+      home: TestScreen(),
 
-      home: MyPageHome(),
 
 
     );
   }
 }
 
-
+//d
