@@ -1,10 +1,11 @@
 import 'package:hankkitoktok/models/base_model.dart';
 import 'package:hankkitoktok/models/meal/meal_delivery.dart';
-import 'package:hankkitoktok/models/meal/ordered_meal.dart';
+import 'package:hankkitoktok/models/meal/meal_delivery_data.dart';
+import 'package:hankkitoktok/models/meal/meal_delivery_order.dart';
 import 'package:hankkitoktok/models/enums.dart';
 
 class Order extends BaseModel{
-  int orderID; //주문번호
+  String orderID; //주문번호
   OrderType orderType; // 주문타입 (일 결제, 주간 결제)
   OrderState? orderState; // 주문상태 (주문완료, 배송중, 배송완료)
   String specialInstruction; //요청사항
@@ -23,10 +24,10 @@ class Order extends BaseModel{
   DateTime? orderTime; // 주문닐짜
 
 
-  late List<MealDelivery> mealDeliveries; // 주문한 도시락 리스트(주간 결제일 경우 2개이상, 일 결제일 경우 1개이상)
+  List<MealDelivery> mealDeliveries = []; // 주문한 도시락 리스트(주간 결제일 경우 2개이상, 일 결제일 경우 1개이상)
 
   Order.init({
-    this.orderID = 0,
+    this.orderID = "",
     this.orderType = OrderType.IMMEDIATE,
     this.orderState = OrderState.ORDERED,
     this.specialInstruction = '',
@@ -39,15 +40,13 @@ class Order extends BaseModel{
     this.orderTime,
     this.totalMealDeliveryCount = 0,
     this.remainingMealDeliveryCount = 0,
-  }){
-    mealDeliveries = [];
-  }
+  });
 
 
   @override
   Order fromMap(Map<String, dynamic> map) {
     return Order.init(
-      orderID: _extractOrderId(map['orderId']),  // OrderId 객체 처리
+      orderID: map['orderId'],  // OrderId 객체 처리
       orderType: _getOrderType(map['orderType']),
       orderState: _getOrderState(map['orderState']),
       specialInstruction: map['specialInstruction'] ?? '',  // null 처리
@@ -58,17 +57,6 @@ class Order extends BaseModel{
       totalPrice: map['orderPrice']?['totalPrice']?['amount'] ?? 0,
       orderTime: map['orderTime'] != null ? DateTime.parse(map['orderTime']) : null,
     );
-  }
-
-  int _extractOrderId(dynamic orderIdObj) {
-    // OrderId 객체에서 값을 추출하는 로직
-    if (orderIdObj is String) {
-      // 혹시 orderId가 단순 문자열이라면 파싱
-      return int.tryParse(orderIdObj) ?? 0;
-    }
-    // 다른 형태로 받는다면 적절한 변환 로직 추가
-    // 기본값으로 0 반환
-    return 0;
   }
 
   OrderType _getOrderType(String? orderTypeStr) {
@@ -108,13 +96,21 @@ class Order extends BaseModel{
     };
   }
 
+  Future<void> setMealDeliveries() async {
+    Map<String, dynamic> query = {'orderId': orderID};
+    mealDeliveries = await networkGetDeliveryList(query, DeliveryListRequestMode.BY_ORDER_ID);
 
+    for(var mealDelivery in mealDeliveries) {
+      print("mealDeliveryId: ${mealDelivery.mealId}");
+      await mealDelivery.setMeal();
+    }
+  }
 
 
   List<MealDetail> get combinedMenuList {
     List<MealDetail> mealDetails = [];
     for (var item in mealDeliveries) {
-      mealDetails.add(item.orderedMeal.getMealDetail());
+      mealDetails.add(item.getMealDetail());
     }
 
     mealDetails.sort((a, b) => a.priority.compareTo(b.priority));
@@ -138,6 +134,9 @@ class Order extends BaseModel{
 
   //2024년 06월 29일 오후 14:20
   String get orderDateString {
+    if(orderTime == null){
+      return '';
+    }
     return '${orderTime!.year}년 ${orderTime!.month}월 ${orderTime!.day}일 ${orderTime!.hour}:${orderTime!.minute}';
   }
   //2024.06.29
