@@ -6,7 +6,7 @@ import 'package:hankkitoktok/functions/httpRequest.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:http/http.dart' as http;
 
-Future<List<Order>> orderGetList(Map<String,dynamic>? query) async {
+Future<List<Order>> orderGetList(Map<String,dynamic> query) async {
   SharedPreferences prefs = await SharedPreferences.getInstance(); // 저장소
   String accessToken = prefs.getString('access_token') ?? '';
   Uri uri = Uri.parse('$BASE_URL/api/v1/orders');
@@ -41,7 +41,6 @@ Future<List<Order>> orderGetList(Map<String,dynamic>? query) async {
     if(response.statusCode == 200){
       var responseBody = jsonDecode(utf8.decode(response.bodyBytes));
       List<Order> result = [];
-      debugPrint(responseBody['result']['content'].toString());
       for (var data in responseBody['result']['content']) {
         Order order = Order.init();
         result.add(order.fromMap(data));
@@ -99,4 +98,51 @@ Future<String> orderPost(Map<String,dynamic> body) async {
     debugPrint(e.toString());
     throw Exception("네트워크 요청에 실패했습니다: $e");
   }
+}
+//Todo: order문제
+Future<Order> networkGetOrder(String orderId) async
+{
+  SharedPreferences prefs = await SharedPreferences.getInstance(); // 저장소
+  String accessToken = prefs.getString('access_token') ?? '';
+  Uri uri = Uri.parse('$BASE_URL/api/v1/orders/$orderId');
+
+  print("$BASE_URL/api/v1/orders/$orderId");
+
+  http.Response? response;
+  Map<String, String> header = {
+    'Content-Type': 'application/json',
+    //'Authorization': 'Bearer $accessToken',
+    'Access-token': 'Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxIiwiaWF0IjoxNzIzNDYyOTM5LCJleHAiOjE3NDkzODI5Mzl9.rmDSuxTSfjJplWLm-v1AxKrz_-9jt8u5RJeC4q2JW38'
+  };
+  try {
+    response = await http.get(uri, headers: header);
+    if (response == null) {
+      throw Exception('리스폰스가 null입니다.');
+    }
+
+    if (response.statusCode == 401) {
+      // access token이 만료되었을 경우,
+      await tokenRefresh(prefs); // refresh token으로 token을 refresh한 후 다시 요청
+      header['Authorization'] = 'Bearer ${prefs.getString('access_token')}';
+      response = await http.get(uri, headers: header);
+
+      if (response == null) {
+        throw Exception('리스폰스가 null입니다.');
+      }
+    }
+    if(response.statusCode == 200){
+      var responseBody = jsonDecode(utf8.decode(response.bodyBytes));
+      Order order = Order.init().fromMap(responseBody['result']);
+      return order;
+      // await prefs.setString("access_token", responseBody['access']); //Todo: 데이터 보고 교체
+      // await prefs.setString("refresh_token", responseBody['refresh']); //Todo: 데이터 보고 교체
+
+    } else {
+      throw Exception(response.statusCode.toString());
+    }
+  }catch (e) {
+    debugPrint(e.toString());
+    throw Exception("네트워크 요청에 실패했습니다: $e");
+  }
+
 }
