@@ -3,20 +3,21 @@ import 'package:hankkitoktok/component/four_image.dart';
 import 'package:hankkitoktok/component/order_detail.dart';
 import 'package:hankkitoktok/const/style.dart';
 import 'package:hankkitoktok/controller/tmpdata.dart';
+import 'package:hankkitoktok/controller/user_controller.dart';
 import 'package:hankkitoktok/functions/httpRequest.dart';
 import 'package:hankkitoktok/models/order/order.dart';
+import 'package:hankkitoktok/models/order/order_data.dart';
 import 'package:intl/intl.dart';
 import 'package:hankkitoktok/models/enums.dart';
 import '../../component/price_info.dart';
 import '../../const/color.dart';
 import '../../const/style2.dart';
-import '../../functions/formatter.dart';
-import '../../models/meal/ordered_meal.dart';
-import '../../models/order/order_user.dart';
+import 'package:get/get.dart';
 import '../../models/user/user.dart';
+import '../../models/user/user_data.dart';
 
 class OrderHistoryDetailScreen extends StatefulWidget {
-  final int orderId;
+  final String orderId;
 
   final formatter = NumberFormat('#,###');
 
@@ -24,35 +25,22 @@ class OrderHistoryDetailScreen extends StatefulWidget {
 
   @override
   State<OrderHistoryDetailScreen> createState() =>
-      _OrderHistoryDetailScreen1State();
+      _OrderHistoryDetailScreenState();
 }
 
-class _OrderHistoryDetailScreen1State extends State<OrderHistoryDetailScreen> {
+class _OrderHistoryDetailScreenState extends State<OrderHistoryDetailScreen> {
 
   late OrderState orderState;
-  late Order order;
-  late User user;
-  void getOrderData() {
-    setState(() {
-      // order = await networkGetRequest(
-      //     model, "/api/v1/orders/${widget.orderId}", null);
-      order = Order.init();
-      for(var item in orders){
-        if(item.orderID == widget.orderId){
-          order = item;
-        }
-      }
-      orderState = order.orderState!;
-
-    });
+  Order order = Order.init();
+  User user = User.init();
+  void getOrderData() async{
+    order = await networkGetOrder(widget.orderId);
+    await order.setMealDeliveries();
+    //await order.setMealDeliveries();
+    orderState = order.orderState!;
     int userId = order.userId;
-    setState(() {
-      user = exampleUser; //userId 가 userId인 유저
-      //   user = await networkGetRequest(
-      //       model, "/api/v1/user/$userId", null);
-      // });
-    });
-
+    //user = (await networkGetUser(userId))!;
+    setState(() {});
   }
 
   @override
